pyproject.toml
[project]
name = "azureml-client"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-ml>=1.32.0",
    "azure-identity>=1.25.3",
    "azure-storage-file-share>=12.17.0",
    "azure-storage-blob>=12.20.0",
    "click>=8.1.0",
]

[project.scripts]
sync_notebooks = "sync_notebooks:cli"

[tool.ruff.lint]
ignore = ["T201", "S106"]
